rename(util): 유틸리티 클래스들을 util 디렉토리로 이동
- 디렉토리 구조 파악을 쉽게 하기 위해 디렉토리 구조를 변경하였다.
- 프로덕션 코드의 위치 수정에 따른 테스트 코드들의 위치도 수정하였다.
