+++ src/main/java/racingcar/util/NumericUtility.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
