+++ src/main/java/racingcar/util/OutputHandler.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import java.util.List;
 import racingcar.domain.Car;
