+++ src/main/java/racingcar/util/Validator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import java.util.stream.Stream;
