+++ src/test/java/racingcar/domain/JudgeTest.java
@@ -9,7 +9,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.InputHandler;
+import racingcar.util.InputHandler;
 
 class JudgeTest {
     private Judge judge;
