+++ src/test/java/racingcar/util/InputHandlerTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,6 +8,7 @@ import java.io.InputStream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.util.InputHandler;
 
 class InputHandlerTest {
     private static final InputStream originalSystemIn = System.in;
