+++ src/test/java/racingcar/util/ValidatorTest.java
@@ -1,9 +1,10 @@
-package racingcar;
+package racingcar.util;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.util.Validator;
 
 class ValidatorTest {
